# perf(EMI-1685): re-enable slow test