The type of this PR is: **Perf**

<!-- Build / Chore / CI / Docs / Feat / Fix / Perf / Refactor / Revert / Style / Test -->

<!-- If applicable, write the Jira ticket number in square brackets e.g. `[PROJECT-XX]`
     The Jira integration will turn it into a clickable link for you. -->

This PR relates to [EMI-1685]
cc @artsy/emerald-devs 

### Description
**Note: after enabling the test it immediately failed due to a timeout. I therefore increased the timeout to make the test pass. Open to feedback on the approach, and we could check out the intermediate commit if we want to try more improvements.**

This PR makes refinements to some of our test helpers and re-enables a test that has been skipped. The test has been ported from enzyme and hopefully performance will be better now.

The intermediate commit 330b3539251c69f9090124ac4d99a233127845ce shows two different implementations of the once-problematic test for comparison. In short the improvements in this PR don't make a meaningful difference in runtime. Repeated runs in that commit show roughly the same timing, with the unoptimized initial code seeming to win slightly more often (difference usually ~100ms, but sometimes the newer implementation wins). The advantage is that we spend less time waiting for arbitrary `flushPromiseQueue()`s to resolve. One possible explanation for this difference is that the calls to `act()` around address form filling helpers may be adding a few ms. Otherwise the time spent waiting for the `flushPromiseQueue()` after filling in the form seems to mostly move into the form-filling time with `act()`.

On my M1 mac this re-enabled test usually took about 5100 ms. However depending on the load of the system it could sometime climb up past 8000ms. This sensitivity has been worse in CI, which is why we disabled the spec to begin with.

Longer but still pretty naive analysis of the different optimizations i have tried [in slack here](https://artsy.slack.com/archives/C2TQ4PT8R/p1733252086261459?thread_ts=1733245233.848769&cid=C2TQ4PT8R) 🔒. I noticed the biggest improvement (~1600ms) by stubbing out the palette Input in tests with a very basic html input/label, but this wouldn't be good for testing purposes.

Related changes
- In [this PR](https://github.com/artsy/force/commit/78d40839a9dd9699b2e717dbaefbdf4c84bf2369#diff-33851a69a0b6c5b35f9e2f3b152c1792af02bbb718ac49e2036f555a7c74b273R218) I changed the test file name, and from looking at the diff on this page I can't see where the original skipped test was added. The
- related test was disabled in https://github.com/artsy/force/pull/13435

[EMI-1685]: https://artsyproduct.atlassian.net/browse/EMI-1685?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ