@erikdstock 2024-12-04
😠 the results are in and if merged [this will become the slowest test](https://app.circleci.com/insights/github/artsy/force/workflows/default/tests?branch=erik.emi-1685-test-speed) at 12s. We could go back to the drawing board and try to improve the speed, give up and just disable the test, or change it to be about editing an address instead so there is less to fill in.